Fixed a few warnings.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@4381 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/app/MessageBody.h
@@ -215,7 +215,7 @@ status_t BMessageBody::ReplaceData(const char *name, int32 index,
 		{
 			debugger("\n\n\tyou \033[44;1;37mB\033[41;1;37me\033[m screwed\n\n");
 		}
-		if (index < RItem->Data().Size())
+		if (index < (int32)RItem->Data().Size())
 		{
 			RItem->Data()[index] = data;
 		}
